openapi-clients/python/demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "holy-bible-api-python >= 7.0.0",
]
